Rename utility to avoid naming collisions
