+++ src/index.js
@@ -84,7 +84,7 @@ export {
 
 // new
 export { default as variant } from './variant'
-export { default as css } from './css'
+export { default as mixed } from './mixed'
 
 // aliases for v2 api
 export { style as responsiveStyle } from './util'
